implemented load from terminal with two arguments. A lot of refactoring. Now object is not depend on constants.
