+++ MazeController.py
@@ -1,5 +1,5 @@
 from Position import Position
-from typing import Dict, Tuple, List
+from typing import Dict
 from Direction import Direction
 
 
@@ -21,11 +21,6 @@ class MazeNode:
 
 class MazeController:
     def __init__(self):
-        self.numpy_maze = []
-        self.cookie_spaces = []
-        self.powerup_spaces = []
-        self.reachable_spaces = []
-        self.ghost_spawns = []
         self.nodes: Dict[Position, MazeNode] = {}
 
     def add_node(self, position: Position):
@@ -40,23 +35,23 @@ class MazeController:
 
     def read_maze(self, maze):
 
-        #створити вузли MazeNode
+        # створити вузли MazeNode
         for y, row in enumerate(maze):
             for x, cell in enumerate(row):
                 position = Position(x, y)
                 self.add_node(position)
 
-        #Заповнити вузли сусідами
+        # Заповнити вузли сусідами
         for y, row in enumerate(maze):
             for x, cell in enumerate(row):
                 position = Position(x, y)
-                #додати телепорти
+                # додати телепорти
                 if x == 0:
                     self.add_edge(position, Position(len(row)-1, y), Direction.LEFT)
                 if x == len(row) - 1:
                     self.add_edge(position, Position(0, y), Direction.RIGHT)
 
-                #звичайні напрями вліво\вправо\вгору\вниз
+                # звичайні напрями вліво\вправо\вгору\вниз
                 if x > 0 and row[x - 1] != 'X':
                     self.add_edge(position, Position(x - 1, y), Direction.LEFT)
                 if x < len(row) - 1 and row[x + 1] != 'X':
